Change doubly-linked_list cpp file to header file
Initially, the whole cpp file was divided into header and implementation file as usual. After separation, it doesn't build as linker connot link methods' declaration and implementation. It is because of template. One possible solution would be explicitly declare datatype of class at a time of using it; however, it is no longer truly templated. Therefore, another solution is taken: include implementation in header file.