Update scheduling.cpp
The container of Job is changed from std::vector to std::list. The main process of job scheduling is about sorting and removing elements. While time complexity of sorting on both structure is the same as O(NlogN) where N is the number of elements, time complexity of erase() is not the same. std::list's erase() methods takes O(N) where N is the number of elements; however, std::vector's takes O(N + M) with M as the number of elements after the last element delete for moving.