Initial commit for mst.cpp
This is an algorihtm that makes an undirected-weighted graph be a minimum spanning tree, which is a completely connected tree with minimum total weight of used edges. Now, the work is not done; class Graph itself with struct Edge and required methods of Graph for mst algorithm are complete. minimum_spanning_tree() is in progress.