Initial commit for graph_coloring.cpp
Graph coloring is an algorithm that decides color of vertices of given graph with adjacent vertices not having the same color. This algorithm greedily picks a color of a vertex based on its adjacent vertices' color; it simply picks colors that are not used by adjacent vetices yet until all vertices' color is decided.