Initial commit for BOJ1707: 이분 그래프
Given a graph, it checks if the graph is bipartite or not. First, using DFS algorithm(recursive version), it colors the nodes with two colors. Specifically, a color is used for a node, and then the other color is used for its adjacent nodes. Then, with the colored graph, it checks whether the adjacent nodes have the same color.