Initial commit for BOJ12946: 육각 보드
Given slots to color in a hexagonal board, it gives the number of colors to color all given slots. The adjacent slots cannot have the same color. In general, the graph coloring problem is NP-complete, which means that it can be solved only with the brute-force manner. However, in this case of hexagonal board, the possible solutions are 0, 1, 2, or 3. If there is no valid slot, the number of color is 0 obviously. If there are only slots that all don't have adjancet slots, the answer is 1. Choosing between 2 and 3 can be done by using bipartite algorithm. If all the given connected components are bipartite, the answer is 2; otherwise, the answer is 3. This idea is obtainned by observation.